fix : mock data 분리
