+++ front/src/__test__/components/Slider.test.tsx
@@ -1,6 +1,6 @@
 import 'jsdom-global/register'; 
 import React from "react";
-import { MovieProp } from '../testUtils';
+import { MovieProp } from './Mock.data';
 import { mount } from 'enzyme';
 import Slider from '../../components/Slider';
