+++ front/src/__test__/components/TextEditor.test.tsx
@@ -1,6 +1,6 @@
 import 'jsdom-global/register'; 
 import React from "react";
-import { ReviewProp } from '../testUtils';
+import { ReviewProp } from './Mock.data';
 import { useSelector } from 'react-redux';
 import { mount } from 'enzyme';
 import TextEditor from '../../components/Review/TextEditor';
